Fix multiblock being dark
The reason this happened was because it was rendering with the relative pos in relation to the world origin instead of relative to the block entity. However, instead of fixing that, I set it to use a position outside the world so that the sides neighboring other blocks isn't affecting the lighting on any of the sides of the hologram blocks. Unfortunately, the problem with this is only sunlight affects the hologram. Better than it was before though.